backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Mujtaba <mujtaba@tripledsystems.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework-jwt = "^1.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
